♻️ refactor: useSearchParams 대응을 위해 MainPage를 클라이언트 컴포넌트로 분리
