+++ src/app/page.tsx
@@ -1,38 +1,17 @@
-'use client';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useState } from 'react';
+import { Suspense } from 'react';
 
-import AllActivities from './_components/AllActivities';
-import PopularActivities from './_components/PopularActivities';
-import Search from './_components/Search';
-import SearchResult from './_components/SearchResult';
+import MainPage from './_components/MainPage';
 
-export default function Home() {
-  const [isLoading, setIsLoading] = useState(false);
-
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const keyword = searchParams.get('search') || '';
-
-  const handleSearch = (query: string) => {
-    if (query.trim() === '') return;
-    const encoded = encodeURIComponent(query);
-    router.push(`/?search=${encoded}`);
-  };
-
+export default function HomePage() {
   return (
-    <div className='mx-auto max-w-1200'>
-      <div className='mx-24 flex min-h-800 flex-col gap-40 pb-136 md:mx-30 md:gap-80 md:pb-204 lg:mx-40 lg:pb-218'>
-        <Search isLoading={isLoading} onSearch={handleSearch} />
-        {keyword ? (
-          <SearchResult keyword={keyword} onLoadingChange={setIsLoading} />
-        ) : (
-          <>
-            <PopularActivities />
-            <AllActivities />
-          </>
-        )}
-      </div>
-    </div>
+    <Suspense
+      fallback={
+        <div className='flex h-200 items-center justify-center'>
+          <div className='border-primary-500 size-50 animate-spin rounded-full border-2 border-t-transparent' />
+        </div>
+      }
+    >
+      <MainPage />
+    </Suspense>
   );
 }
